requirements.txt
boto3>=1.4.5
configparser
PyYAML
tabulate
win-unicode-console
